Python macro file management
Some issues were found when attempting to use the Kivy GUI functions with Pynput listeners.  Also the recorder stops recording outside of the Kivy GUI if it is not the current application when you use Kivy mouse/keyboard listeners.  It is clear the recording will probably use this file with Pynput in a background process in the future.  Most of the main.py program has been moved to this file (macrofilecontroller.py) to complete the file creation, editing, recording, saving and loading of python macro files.  This file will later be incorporated with either a Kivy GUI as a separate process or maybe a Pygame GUI.  This file has testing functions near the bottom of the document that should be either removed or edited in the future.